Filter events in "next events" list
Most people just want to see when the next
(Open|Self|Rad|Zerspanungs|...)Lab takes place. They are not interested
in internal events like "OrgaTreffen". So we have to filter out such
events from the "Nächste Termine" list on the start page.

The easiest (but dirtiest) way is to do this with javascript :tada: